+++ src/main/java/com/comphenix/protocol/utility/ByteBuddyGenerated.java
@@ -1,0 +1,9 @@
+package com.comphenix.protocol.utility;
+
+/**
+ * Represents an object that has been generated using ByteBuddy.
+ *
+ * @author Pim
+ */
+public interface ByteBuddyGenerated {
+}
